Trying to fix stack frame issue
