+++ api/Dockerfile
@@ -1,5 +1,5 @@
 FROM openjdk:12
 WORKDIR /usr/src/ab2d-api
 ADD target /usr/src/ab2d-api
-CMD java -jar api-*-SNAPSHOT.jar
+CMD java -jar -Xverify:none api-*-SNAPSHOT.jar
 EXPOSE 8080
