+++ api/src/main/java/gov/cms/ab2d/api/SpringBootApp.java
@@ -3,9 +3,11 @@ package gov.cms.ab2d.api;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.domain.EntityScan;
+import org.springframework.context.annotation.ComponentScan;
 
 @SpringBootApplication
 @EntityScan(basePackages = {"gov.cms.ab2d.domain"})
+@ComponentScan(basePackages = {"gov.cms.ab2d.domain", "gov.cms.ab2d.api"})
 public class SpringBootApp {
 
     public static void main(String[] args) {
