+++ pom.xml
@@ -34,7 +34,6 @@
             <version>1.18.10</version>
             <scope>provided</scope>
         </dependency>
-
     </dependencies>
 
     <build>
